refactor(tauri): split git command module
