+++ apps/desktop/src-tauri/src/commands/git/mod.rs
@@ -1,0 +1,14 @@
+mod authorization;
+mod command_handlers;
+mod snapshot;
+
+pub(crate) use authorization::invalidate_worktree_resolution_cache_for_repo;
+pub use command_handlers::{
+    git_commit_all, git_commits_ahead_behind, git_create_worktree, git_get_branches,
+    git_get_current_branch, git_get_diff, git_get_status, git_get_worktree_status,
+    git_get_worktree_status_summary, git_pull_branch, git_push_branch, git_rebase_abort,
+    git_rebase_branch, git_remove_worktree, git_switch_branch,
+};
+
+#[cfg(test)]
+mod tests;
