+++ apps/desktop/src-tauri/src/commands/git/tests/mod.rs
@@ -1,0 +1,4 @@
+mod authorization;
+mod command_handlers;
+mod fixtures;
+mod snapshot;
